requirements.txt
pyrealsense2
numpy
pyyaml==6.0
pathlib
DateTime
pysimplegui